requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-allauth==0.57.0
channels==4.0.0
channels-redis==4.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.4.0
cloudinary==1.40.0
stripe==10.8.0
google-generativeai==0.7.2
celery==5.3.4
redis==5.0.8
drf-yasg==1.21.7
django-filter==23.5
whitenoise==6.6.0
gunicorn==21.2.0
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.0
daphne==4.1.2
requests==2.31.0
python-dotenv==1.0.01
dj-database-url==2.1.0
